[ao] Added DetectorQConfigMapping class framework
Summary: This adds the framework for a DetectorQConfigMapping class that
stores the QConfigMapping generated by a specific detector and also
allows the merging of two of these DetectorQConfigMappings to merge
suggestions from different detectors since the logic interacts in
different ways.

This is meant to be used in the ModelReport class to be able to get
suggestions from each of the detectors in the report and merge them
together to return a single merged mapping suggestion.

Test Plan: python test/test_quantization.py TestFxDetectorQConfigMapping

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: bf930c92e93790d115165b7d847c3ae2334dae5c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/83454