Add specialized lerp kernel to accelerate `make_tensor`
`make_tensor` used several times for every single opinfo test, so small gains in
performace will result in significant improvements in test completion time.
Aside from the random number generation, the main limitation for floating an
complex dtypes is the scaling operation from [0, 1] to [low, high]. This
operation is a linear interpolation (`lerp`) but calling `torch.lerp` is
actually not faster because the kernel isn't optimised for scalar `low` and
`high` values. Particularly on CUDA, this means extra kernel launches and device
synchronization to copy the values into tensors.

This adds a new `_lerp_scalar` kernel for use in `make_tensor` and also reduces
the number of seperate kernels launched for complex tensors, since for the small
tensors used in tests it is often the kernel launch overhead that is the
limiting factor.

Here are benchmark results for `make_tensor` with small and medium sized tensors.

| Device | dtype     | size | Master (us) | This PR (us) | Speedup |
|--------|-----------|------|-------------|--------------|---------|
| cpu    | float32   | 8    | 19.6        | 8.61         | 2.3     |
|        |           | 4096 | 34.9        | 23.3         | 1.5     |
|        | complex64 | 8    | 37.4        | 10.1         | 2.7     |
|        |           | 4096 | 70.3        | 39.2         | 1.8     |
| cuda   | float32   | 8    | 38.7        | 28.2         | 1.4     |
|        |           | 4096 | 36.2        | 18.3         | 1.4     |
|        | complex64 | 8    | 74.9        | 20.6         | 3.6     |
|        |           | 4096 | 74.1        | 20.8         | 3.6     |

ghstack-source-id: 26f756475e25c2a0d2b27279d92c1b2a62d6b0bd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84846