Remove eager mode support form CommTensor
We don't need eager mode support (automatic wait on read) for now.
Removing that to simply the code. We can always add this back if
necessary in the future.

Note that, we still need the eager mode code in `__torch_dispatch__`,
as `make_fx` will also run the ops in eager mode to get the output.

ghstack-source-id: fd82b42585f5f46527ced3cac08fbb6a42f1a1bc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84978